+++ actors/bundler/CHANGELOG.md
@@ -51,7 +51,7 @@
 - Change `docker-compose.yml` to have redis `volumes: ./redis-data:/data` to `volumes: /redis-data` (volumes finally mount on docker env instead of host machine)
 - update imports with SDK renames (see SDK changelog)
 - use `--max-latency` option for batcher in `docker-compose.yml`
-- move `OperationStatus` enum to `@nocturne-xyz/sdk`
+- move `OperationStatus` enum to `@nocturne-xyz/core`
 - use `node:18.12.1` in dockerfile
 - Add docker internal host gateway network to server and submitter so they can connect to hh
 - Expose server ports in docker compose file